Add plot rules in the workflow
